requirements.txt
Django==3.2.10
djangorestframework==3.12.4
psycopg2-binary==2.9.1
requests==2.26.0
pandas==1.3.3
numpy==1.21.2
matplotlib==3.4.3
reportlab==3.6.1
gunicorn==20.1.0
scikit-learn==0.24.2
python-dotenv==0.19.0
dj-database-url==0.5.0